Redesign of movman to atomically schedule with overwriting
Queue now represents a single composite movement, with one priority
associated to it. Moves are overwritten only when a higher priority
movement is requested. This will resolve a lot of annoying issues caused
by the "feature" (which I never actually used) of doing delayed
scheduling of low priority moves. It was useless.